requirements-dev.txt
pytest
ddbmock
